# mec-exposure/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: MEC Exposure & Experience Management API
  description: "# Introduction\n---\n\nUse the MEC Exposure & Experience Management\
    \ API for discovery and utilisation of Multi-Access Edge Compute (MEC) \nresources\
    \ and services. \n\n\n---\n# API Scope\n\n---\n\nAPIs defined in this version\
    \ of the specification can be categorized into the following areas:\n\n* __MEC\
    \ Edge Discovery__ - discovery of optimal MEC platforms\n* __MEC Service Discovery__\
    \ - discovery of optimal Service Endpoints for client connections\n* __MEC Service\
    \ Profile management__ - register, retrieve, update, and delete MEC Service Profiles\n\
    * __MEC Service Endpoint management__ - register one or more Service Endpoints,\
    \ and retrieve, update, and delete those registrations\n* __MEC Edge Event notification__\
    \ - subscription management for MEC and Service discovery Event notifications\n\
    \n---\n# Definitions\n---\nThis section provides definitions of terminologies\
    \ commonly referred to throughout the API descriptions.\n\n* __API Consumer__\
    \ - A user or an application consuming the MEC Exposure and Experience Management\
    \ APIs. \n* __Application Service Provider__ - the developer/publisher who deploys\
    \ applications on MEC platforms.\n* __Density__ - Minimum 4G/5G subscriber density\
    \ in a geographical area, represented as the number of subscribers per square\
...
  license:
    name: Apache-2.0
...
servers:
...
security:
- oAuth2:
  - edge:discovery:read
  - edge:serviceprofile:read
  - edge:serviceprofile:write
  - edge:serviceregistry:read
  - edge:serviceregistry:write
...
paths:
...
  /serviceendpoints:
    get:
      tags:
      - Discovery
      summary: Find optimal Service Endpoints for clients to connect to
      description: "Returns a list of optimal Service Endpoints that client devices\
        \ can connect to. You can search based on Service Profile, Region, subscriber\
        \ density or UEIdentity."
      operationId: get_service_endpoints
      parameters:
      - name: region
        in: query
        description: MEC region ID
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: zone
        in: query
        description: MEC zone ID
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: serviceEndpointsIds
        in: query
        description: "serviceEndpointsIds, delimited by pipe. Each serviceEndpointsId\
          \ is a pointer to the service endpoints (IP address, FQDN etc.) previously\
          \ registered for a given service."
        required: true
        style: pipeDelimited
        explode: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/types_serviceEndpoints_Id'
      - name: subscriberDensity
        in: query
        description: Minimum number of 4G/5G subscribers per square kilometer.
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: UEIdentityType
        in: query
        description: Type of User Equipment identifier used in `UEIdentity`.
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/types_UEIdentityType'
      - name: UEIdentity
        in: query
        description: Identifier value for User Equipment. The type of identifier is
          defined by the UEIdentityType parameter.
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/types_UEIdentity'
      responses:
        "200":
          description: Find optimal Service Endpoints for clients to connect to
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetServiceEndpointsResponse'
        "401":
          description: HTTP 401 Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        default:
          description: HTTP 500 Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
      x-openapi-router-controller: swagger_server.controllers.discovery_controller
    post:
      tags:
      - Service Endpoints
      summary: Register Service Endpoints
      description: Register Service Endpoints of a deployed application to specified
        MEC Platforms.
      operationId: register_service_endpoints
      requestBody:
        description: Array of Service Endpoints for a deployed application
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/resources_edge_hosted_service'
      responses:
        "201":
          description: Returns a serviceEndpointsId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostServiceEndpointResponse'
        "401":
          description: HTTP 401 Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        default:
          description: HTTP 500 Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
      x-openapi-router-controller: swagger_server.controllers.service_endpoints_controller
  /serviceendpoints/{serviceEndpointsId}:
    get:
      tags:
      - Service Endpoints
      summary: Get registered edge service endpoint information
      description: Returns endpoint information for all Service Endpoints registered
        to a specified serviceEndpointId.
      operationId: get_service_endpoints_by_service_endpoints_id
      parameters:
      - name: serviceEndpointsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/types_serviceEndpoints_Id'
      responses:
        "200":
          description: Arry of registered service endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/resources_edge_hosted_service'
                x-content-type: application/json
        "400":
          description: HTTP 400 Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        "401":
          description: HTTP 401 Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        "404":
          description: HTTP 404 The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        default:
          description: HTTP 500 Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
      x-openapi-router-controller: swagger_server.controllers.service_endpoints_controller
    put:
      tags:
      - Service Endpoints
      summary: Update a Service Endpoint
      description: Update registered Service Endpoint information.
      operationId: update_service_endpoint
      parameters:
      - name: serviceEndpointsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/types_serviceEndpoints_Id'
      requestBody:
        description: Service Endpoint information
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/resources_edge_hosted_service'
      responses:
        "204":
          description: HTTP 204 No Content
        "401":
          description: HTTP 401 Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        "404":
          description: HTTP 404 The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        default:
          description: HTTP 500 Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
      x-openapi-router-controller: swagger_server.controllers.service_endpoints_controller
    delete:
      tags:
      - Service Endpoints
      summary: Deregister an application's Service Endpoint
      description: Deregister an application's Service Endpoint from the MEC Platform(s).
      operationId: deregister_service_endpoint
      parameters:
      - name: serviceEndpointsId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/types_serviceEndpoints_Id'
      responses:
        "204":
          description: HTTP 204 No Content
        "400":
          description: HTTP 400 Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        "401":
          description: HTTP 401 Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        "404":
          description: HTTP 404 The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
        default:
          description: HTTP 500 Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types_error'
      x-openapi-router-controller: swagger_server.controllers.service_endpoints_controller
...
components:
  schemas:
...
    resources_edge_hosted_service:
      type: object
      properties:
        ern:
          $ref: '#/components/schemas/types_edgeResource'
        serviceEndpointsId:
          $ref: '#/components/schemas/types_serviceEndpoints_Id'
        serviceEndpoint:
          $ref: '#/components/schemas/resources_service_endpoint'
        applicationServerProviderId:
          type: string
          description: Unique ID representing the Edge Application Provider
        applicationId:
          type: string
          description: Unique ID representing the Edge Application
        serviceDescription:
          type: object
          additionalProperties: false
          description: Description of the Service Endpoint
        serviceProfileID:
          $ref: '#/components/schemas/types_service_profile_Id'
      description: Edge hosted service represented by Service Endpoint definition
      example:
        ern: {}
        serviceEndpointsId: {}
        applicationServerProviderId: applicationServerProviderId
        serviceProfileID: {}
        serviceDescription: {}
        serviceEndpoint:
          FQDN: FQDN
          port: 0
          IPv6Address: IPv6Address
          URI: http://example.com/aeiou
          IPv4Address: IPv4Address
        applicationId: applicationId
    resources_service_endpoint:
      type: object
      properties:
        URI:
          type: string
          description: URI of Service Endpoint if available
          format: uri
        FQDN:
          type: string
          description: FQDN of Service Endpoint if available
        IPv4Address:
          type: string
          description: IPv4 Address of Service Endpoint if available
          format: ipv4
        IPv6Address:
          type: string
          description: IPv6 Address of Service Endpoint if available
          format: ipv6
        port:
          type: integer
          description: |
            Port information of Service Endpoint if IPv4 or IPv6 is mentioned
      additionalProperties: false
      example:
        FQDN: FQDN
        port: 0
        IPv6Address: IPv6Address
        URI: http://example.com/aeiou
        IPv4Address: IPv4Address
...
    types_edgeResource:
      type: object
      additionalProperties: false
      description: |
        An object defined by the service provider representing an edge resource within its network domain, such as a MEC Platform.  The string contains colon-separated metadata in the form   ern:<tsp-identifier>:<tsp-region>:<resource-type>:<resource-identifier>:<tsp-extensions>
...
    types_serviceEndpoints_Id:
      type: object
      additionalProperties: false
      description: |
        A system-defined string identifier representing one or more registered Service Endpoints.
      readOnly: true
    types_service_profile_Id:
      type: object
      additionalProperties: false
      description: |
        Unique identifier for a service profile
      readOnly: false
...
    types_error:
      required:
      - code
      - links
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        links:
          $ref: '#/components/schemas/links_Common'
      additionalProperties: false
...
    types_UEIdentityType:
      type: string
      description: Type of User Equipment identifier used in `UEIdentity`.
      enum:
      - IPAddress
      - MSISDN
      - IMEI
      - MDN
      - GPSI
    types_UEIdentity:
      type: string
      description: Identifier value for User Equipment. The type of identifier is
        defined by the 'UEIdentityType' parameter.
...
    links_Register-service-endpoint:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_Registerserviceendpoint_link'
      additionalProperties: false
      readOnly: true
    links_Update-service-endpoint:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_Updateserviceendpoint_link'
      additionalProperties: false
      readOnly: true
    links_Delete-service-endpoint:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_Deleteserviceendpoint_link'
      additionalProperties: false
      readOnly: true
    links_Optimal-service-endpoints-by-service-profile:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_Optimalserviceendpointsbyserviceprofile_link'
      additionalProperties: false
      readOnly: true
...
    links_Self:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_Self_link'
      additionalProperties: false
      readOnly: true
    links_Version-History:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_VersionHistory_link'
      additionalProperties: false
      readOnly: true
    links_Terms-of-Service:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_TermsofService_link'
      additionalProperties: false
      readOnly: true
    links_Bookmark:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/links_Bookmark_link'
      additionalProperties: false
      readOnly: true
    links_Common:
      type: array
      readOnly: true
      items:
        oneOf:
        - $ref: '#/components/schemas/links_Self'
        - $ref: '#/components/schemas/links_Version-History'
        - $ref: '#/components/schemas/links_Terms-of-Service'
        - $ref: '#/components/schemas/links_Bookmark'
...
    GetServiceEndpointsResponse:
      title: GetServiceEndpointsResponse
      type: object
      properties:
        serviceEndpoints:
          type: array
          items:
            $ref: '#/components/schemas/resources_edge_hosted_service'
        links:
          type: array
          items:
            oneOf:
            - $ref: '#/components/schemas/links_Register-service-endpoint'
            - $ref: '#/components/schemas/links_Common'
      example:
        links:
        - ""
        - ""
        serviceEndpoints:
        - ern: {}
          serviceEndpointsId: {}
          applicationServerProviderId: applicationServerProviderId
          serviceProfileID: {}
          serviceDescription: {}
          serviceEndpoint:
            FQDN: FQDN
            port: 0
            IPv6Address: IPv6Address
            URI: http://example.com/aeiou
            IPv4Address: IPv4Address
          applicationId: applicationId
        - ern: {}
          serviceEndpointsId: {}
          applicationServerProviderId: applicationServerProviderId
          serviceProfileID: {}
          serviceDescription: {}
          serviceEndpoint:
            FQDN: FQDN
            port: 0
            IPv6Address: IPv6Address
            URI: http://example.com/aeiou
            IPv4Address: IPv4Address
          applicationId: applicationId
    PostServiceEndpointResponse:
      title: PostServiceEndpointResponse
      type: object
      properties:
        serviceEndpointsId:
          $ref: '#/components/schemas/types_serviceEndpoints_Id'
        links:
          type: array
          items:
            oneOf:
            - $ref: '#/components/schemas/links_Update-service-endpoint'
            - $ref: '#/components/schemas/links_Delete-service-endpoint'
            - $ref: '#/components/schemas/links_Optimal-service-endpoints-by-service-profile'
            - $ref: '#/components/schemas/links_Common'
      example:
        serviceEndpointsId: {}
        links:
        - ""
        - ""
...
    links_Registerserviceendpoint_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: RegisterServiceEndpoint
          type: string
          example: RegisterServiceEndpoint
        method:
          pattern: post
          type: string
          example: post
        href:
          pattern: /serviceendpoints
          type: string
          example: /serviceendpoints
    links_Updateserviceendpoint_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: UpdateServiceEndpoint
          type: string
          example: UpdateServiceEndpoint
        method:
          pattern: put
          type: string
          example: put
        href:
          pattern: "\\/serviceendpoints\\/[A-Za-z0-9]*$"
          type: string
          example: /serviceendpoints/98765
    links_Deleteserviceendpoint_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: DeleteServiceEndpoint
          type: string
          example: DeleteServiceEndpoint
        method:
          pattern: delete
          type: string
          example: delete
        href:
          pattern: "\\/serviceendpoints\\/[A-Za-z0-9]*$"
          type: string
          example: /serviceendpoints/98765
    links_Optimalserviceendpointsbyserviceprofile_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: ListServiceEndpointsByServiceProfileId
          type: string
          example: ListServiceEndpointsByServiceProfileId
        method:
          pattern: get
          type: string
          example: get
        href:
          pattern: "\\/serviceendpoints\\?serviceProfileId=[A-Za-z0-9]*$"
          type: string
          example: /serviceendpoints?serviceProfileId=12345
...
    links_Self_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: self
          type: string
          example: self
        method:
          pattern: get
          type: string
          example: get
        href:
          type: string
          format: uri
    links_VersionHistory_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: version-history
          type: string
          example: version-history
        method:
          pattern: get
          type: string
          example: get
        href:
          type: string
          format: uri
    links_TermsofService_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: terms-of-service
          type: string
          example: terms-of-service
        method:
          pattern: get
          type: string
          example: get
        href:
          type: string
          format: uri
    links_Bookmark_link:
      required:
      - href
      - method
      - rel
      type: object
      properties:
        rel:
          pattern: bookmark
          type: string
          example: bookmark
        method:
          pattern: get
          type: string
          example: get
        href:
          type: string
          format: uri
...
  securitySchemes:
    oAuth2:
      type: oauth2
      description: "This API uses OAuth 2 with the implicit, password and client credentials\
        \ grant flow"
      flows:
        implicit:
          authorizationUrl: https://example.com/oauth/authorize
          scopes:
            discovery:read: Grant read-only access to discovery data
            serviceprofile:read: Grant read-only access to service profile data
            serviceprofile:write: Grant write access to service profile data
            serviceregistry:read: Grant read-only access to service registry data
            serviceregistry:write: Grant write access to service registry data
        password:
          tokenUrl: https://example.com/oauth/token
          scopes:
            discovery:read: Grant read-only access to discovery data
            serviceprofile:read: Grant read-only access to service profile data
            serviceprofile:write: Grant write access to service profile data
            serviceregistry:read: Grant read-only access to Service registry data
            serviceregistry:write: Grant write access to Service registry data
        clientCredentials:
          tokenUrl: https://example.com/oauth/token
          scopes:
            discovery:read: Grant read-only access to discovery data
            serviceprofile:read: Grant read-only access to service profile data
            serviceprofile:write: Grant write access to service profile data
            serviceregistry:read: Grant read-only access to Service registry data
            serviceregistry:write: Grant write access to Service registry data
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_oAuth2
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_oAuth2